+++ packages/ragbits-core/CHANGELOG.md
@@ -2,10 +2,10 @@
 
 ## Unreleased
 
-- Image embeddings in PgVectorStore (#502)
-- Add PgVectorStore to vector store integration tests (#502)
+- Image embeddings in PgVectorStore (#495)
+- Add PgVectorStore to vector store integration tests (#495)
 - Add new fusion strategies for the hybrid vector store: RRF and DBSF (#413)
-
+- move sources from ragbits-document-search to ragbits-core (#496)
 
 ## 0.13.0 (2025-04-02)
 - Make the score in VectorStoreResult consistent (always bigger is better)